// src/app/components/shared/model/character.ts
import {Film} from "./film";

export interface Character {
  name: string;
  height: string;
  mass : string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: Film[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: Date;
  edited: Date;
  url: string;
}